refactor: moved services to repos folder, stores - to stores
